function done at school

+++ Cursus42/Libft/ft_bzero.c
@@ -1,7 +1,7 @@
 #include "libft.h"
 void	ft_bzero(void * s, size_t n)
 {
-	int i;
+	size_t i;
 	char *str = (char *)s;
 
 	i = 0;
@@ -11,4 +11,3 @@ void	ft_bzero(void * s, size_t n)
 		i++;
 	}
 }
-//review
